Image file client side validation

+++ Shopy.Core/Application/Products/Delete/DeleteProductCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using Mediator.Net.Context;
 using Mediator.Net.Contracts;
+using Shopy.Core.Exceptions;
 using Shopy.Data;
 using System.Threading;
 using System.Threading.Tasks;
@@ -15,6 +16,12 @@ namespace Shopy.Core.Application.Products.Commands
                 var command = context.Message;
 
                 var toRemove = await dbContext.Products.FindAsync(command.Uid);
+
+                if (toRemove == null)
+                {
+                    throw new ProductNotFoundException(command.Uid);
+                }
+
                 dbContext.Products.Remove(toRemove);
                 await dbContext.SaveChangesAsync();
             }
